Logging + Fixes to MessageQueue
The fix involves a MessageQueueClient's position being removed from the
underlying queue. When this happens, checking whether `mc.position ==
nil` is insufficient. There is no easy way to see if an element was
removed from the queue, so now we explicitly set `el.Value = nil` when
removing that element and we reset a client's position if
    - `mc.position == nil` (ie: no position)
    - `mc.position.Value = nil` (ie: prev position was removed from
      queue)